Seperated animation functions from drawing functions for easier extendability
